@@ -1,21 +1,43 @@
 {
+  "name": "evotm/evosc",
+  "type": "project",
+  "description": "EvoSC is a modern server controller for the game Trackmania² and TM2020.",
+  "keywords": [
+    "Evo",
+    "Maniaplanet",
+    "Trackmania"
+  ],
+  "license": "Creative Commons Attribution-NonCommercial-NoDerivatives 4.0 International Public License",
+  "homepage": "https://github.com/EvoTM/EvoSC",
+  "repositories": [
+    {
+      "type": "vcs",
+      "url": "https://github.com/EvoTM/gbx.git"
+    },
+    {
+      "type": "vcs",
+      "url": "https://github.com/EvoTM/dedicated-server-api.git"
+    }
+  ],
   "require": {
     "php": "^7.4",
     "ext-json": "*",
     "ext-simplexml": "*",
     "ext-mbstring": "*",
     "ext-zip": "*",
-    "maniaplanet/dedicated-server-api": "^5.0",
-    "illuminate/database": "^5.5",
-    "symfony/console": "^4.0",
+    "maniaplanet/dedicated-server-api": "dev-master",
+    "symfony/console": "^5.2",
     "guzzlehttp/guzzle": "^6.3",
     "latte/latte": "^2.4",
-    "illuminate/pagination": "^5.6",
     "symfony/process": "^4.2",
     "symfony/event-dispatcher": "^4.2",
     "composer/ca-bundle": "^1.1",
     "manialib/gbx": "^4.1",
-    "larapack/dd": "^1.1"
+    "larapack/dd": "^1.1",
+    "illuminate/container": "^8.27",
+    "illuminate/database": "^8.27",
+    "mexitek/phpcolors": "^1.0",
+    "league/csv": "^9.7"
   },
   "scripts": {
     "pre-install-cmd": [
@@ -25,7 +47,10 @@
   "autoload": {
     "psr-4": {
       "EvoSC\\": "core/",
-      "EvoSC\\Modules\\": ["core/Modules/", "modules/"]
+      "EvoSC\\Modules\\": [
+        "core/Modules/",
+        "modules/"
+      ]
     }
   },
   "require-dev": {
