@@ -34,7 +34,7 @@ class File
 
             return file_get_contents($fileName);
         } else {
-            Log::error("Could not load file '$fileName'");
+            Log::warning("Could not load file '$fileName'");
         }
 
         return null;
