@@ -2,6 +2,7 @@
 
 namespace EvoSC\Classes;
 
+use Symfony\Component\Console\Formatter\OutputFormatterStyle;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
@@ -71,7 +72,7 @@ class Log
         }
 
         if (isDebug()) {
-            $callingClass .= "\nData: " . json_encode($caller['args']);
+            $callingClass .= "\nData: " . json_encode($caller);
         }
 
         $time = date("H:i:s", time());
@@ -116,10 +117,27 @@ class Log
      * @param      $message
      * @param bool $echo
      */
-    public static function error($message, bool $echo = true)
+    public static function error($message, bool $echo = true, int $limit = 2)
     {
-        list($childClass, $caller) = debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS, 2);
+        list($childClass, $caller) = debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS, $limit);
         self::write('<error>' . $message . '</>', $echo, $caller);
+    }
+
+    /**
+     * Log an error message and cause stacktrace.
+     *
+     * @param      $message
+     * @param      $throwable
+     * @param bool $echo
+     */
+    public static function errorWithCause($message, \Throwable $throwable, bool $echo = true)
+    {
+        self::error($message . ': ' . $throwable->getMessage(), $echo, 3);
+
+        if (isVerbose())
+        {
+            self::write($throwable->getTraceAsString(), $echo);
+        }
     }
 
     /**
@@ -128,10 +146,27 @@ class Log
      * @param      $message
      * @param bool $echo
      */
-    public static function warning($message, bool $echo = true)
+    public static function warning($message, bool $echo = true, $limit = 2)
     {
-        list($childClass, $caller) = debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS, 2);
-        self::write('<fg=red>' . $message . '</>', $echo, $caller);
+        list($childClass, $caller) = debug_backtrace(DEBUG_BACKTRACE_IGNORE_ARGS, $limit);
+        self::write('<warning>' . $message . '</>', $echo, $caller);
+    }
+
+    /**
+     * Log a warning message and cause stacktrace.
+     *
+     * @param      $message
+     * @param      $throwable
+     * @param bool $echo
+     */
+    public static function warningWithCause($message, \Throwable $throwable, bool $echo = true)
+    {
+        self::warning($message . ': ' . $throwable->getMessage(), $echo, 3);
+
+        if (isVeryVerbose())
+        {
+            self::write($throwable->getTraceAsString(), $echo);
+        }
     }
 
     /**
@@ -153,6 +188,9 @@ class Log
     {
         self::$singleFileMode = (bool)config('server.logs.single-file', false);
         self::$logPrefix = (string)config('server.logs.prefix', 'evosc');
+
+        $warningOutputFormatterStyle = new OutputFormatterStyle('red');
+        $output->getFormatter()->setStyle('warning', $warningOutputFormatterStyle);
 
         self::$output = $output;
     }
