@@ -8,7 +8,9 @@ use Maniaplanet\DedicatedServer\InvalidArgumentException;
 use Maniaplanet\DedicatedServer\Structures\PlayerDetailedInfo;
 use Maniaplanet\DedicatedServer\Structures\PlayerInfo;
 use Maniaplanet\DedicatedServer\Structures\PlayerRanking;
+use Maniaplanet\DedicatedServer\Structures\ScriptInfo;
 use Maniaplanet\DedicatedServer\Structures\ServerOptions;
+use Maniaplanet\DedicatedServer\Structures\SystemInfos;
 use Maniaplanet\DedicatedServer\Structures\Version;
 
 /**
@@ -97,7 +99,7 @@ use Maniaplanet\DedicatedServer\Structures\Version;
  * @method static int sendBill(string $string, int $int, string $string, string $string = null)
  * @method static object getBillState(int $int)
  * @method static int getServerPlanets()
- * @method static object getSystemInfo()
+ * @method static SystemInfos getSystemInfo()
  * @method static bool setConnectionRates(int $int, int $int)
  * @method static array getServerTags()
  * @method static bool setServerTag(string $string, string $string)
@@ -132,8 +134,7 @@ use Maniaplanet\DedicatedServer\Structures\Version;
  * @method static string gameDataDirectory()
  * @method static string getMapsDirectory()
  * @method static string getSkinsDirectory()
- * @method static bool setTeamInfo(string $string, double $double, string $string, string $string, double $double, string $string, string $string, double $double, string $string)
- * @method static object getTeamInfo(int $int)
+ * @method static \Maniaplanet\DedicatedServer\Structures\Team getTeamInfo(int $int)
  * @method static bool setForcedClubLinks(string $string, string $string)
  * @method static object getForcedClubLinks()
  * @method static string connectFakePlayer()
@@ -178,7 +179,7 @@ use Maniaplanet\DedicatedServer\Structures\Version;
  * @method static bool getWarmUp()
  * @method static string getModeScriptText()
  * @method static bool setModeScriptText(string $string)
- * @method static object getModeScriptInfo()
+ * @method static ScriptInfo getModeScriptInfo()
  * @method static array getModeScriptSettings()
  * @method static bool setModeScriptSettings(array $modeScriptSettingsArray)
  * @method static bool sendModeScriptCommands(object $struct)
@@ -328,11 +329,27 @@ class Server
         return collect(Server::getMapList())->contains('fileName', $filename);
     }
 
+    public static function callGetScores(): bool
+    {
+        return Server::triggerModeScriptEventArray('Trackmania.GetScores');
+    }
+
+    public static function getModeScriptSetting(string $name)
+    {
+        $data = Server::getModeScriptSettings();
+
+        if (array_key_exists($name, $data)) {
+            return $data[$name];
+        }
+
+        return null;
+    }
+
     /**
      * Call an rpc-method.
      *
-     * @param  string  $rpc_func
-     * @param  null|mixed[]  $args
+     * @param string $rpc_func
+     * @param null|mixed[] $args
      */
     public static function call(string $rpc_func, $args = null)
     {
