@@ -19,14 +19,21 @@ use Symfony\Component\Console\Question\Question;
 
 class AddAdmin extends Command
 {
+    /**
+     * Command settings
+     */
     protected function configure()
     {
         $this->setName('add:admin')
             ->setDescription('Adds player to AdminGroups')
             ->addArgument('login', InputArgument::REQUIRED, 'Player login to add')
-            ->addArgument('groupId', InputArgument::OPTIONAL, 'Group id to add');
+            ->addArgument('groupId', InputArgument::OPTIONAL, 'Group id to add, 1 = Masteradmin');
     }
 
+    /**
+     * @param InputInterface $input
+     * @param OutputInterface $output
+     */
     protected function initialize(InputInterface $input, OutputInterface $output)
     {
         Log::setOutput($output);
@@ -34,19 +41,27 @@ class AddAdmin extends Command
         Database::init();
     }
 
+    /**
+     * @param InputInterface $input
+     * @param OutputInterface $output
+     * @return int
+     */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         $login = $input->getArgument('login');
         $groupId = $input->getArgument('groupId');
         $helper = new QuestionHelper();
 
-
         $player = Player::find($login);
 
         if ($player === null) {
-            $output->writeln("No player found with login '{$login}'.");
-            exit(1);
+            $player = Player::create([
+                'Login'        => $login,
+                'NickName'     => $login,
+                'ubisoft_name' => $login,
+            ]);
         }
+
         if (!$groupId) {
             $table = new Table($output);
             $table->setHeaders(["Id", "Title"]);
@@ -67,7 +82,7 @@ class AddAdmin extends Command
             $groupId = intval($groupId);
         } else {
             $output->writeln("Group id must be numeric and integer.");
-            exit(1);
+            return 1;
         }
 
         $groupName = Group::findOrFail($groupId)->Name;
@@ -88,7 +103,9 @@ class AddAdmin extends Command
                 "You have been added to group {$groupName}, please rejoin to this server.", $login);
 
         } catch (Exception $e) {
-            // silent exception
+            Log::errorWithCause("Failed to send message to chat", $e);
         }
+
+        return 0;
     }
 }
