@@ -24,12 +24,12 @@ use EvoSC\Controllers\CountdownController;
 use EvoSC\Controllers\EventController;
 use EvoSC\Controllers\HookController;
 use EvoSC\Controllers\MapController;
-use EvoSC\Controllers\MatchController;
 use EvoSC\Controllers\MatchSettingsController;
 use EvoSC\Controllers\ModuleController;
 use EvoSC\Controllers\PlanetsController;
 use EvoSC\Controllers\PlayerController;
 use EvoSC\Controllers\QueueController;
+use EvoSC\Controllers\RoyalController;
 use EvoSC\Controllers\SetupController;
 use EvoSC\Controllers\TemplateController;
 use EvoSC\Models\AccessRight;
@@ -38,15 +38,23 @@ use EvoSC\Modules\InputSetup\InputSetup;
 use EvoSC\Modules\QuickButtons\QuickButtons;
 use Exception;
 use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Command\SignalableCommandInterface;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
-class EscRun extends Command
+class EscRun extends Command implements SignalableCommandInterface
 {
+    private static bool $docker = false;
+    private bool $keepRunning = true;
+
+    /**
+     * Command settings
+     */
     protected function configure()
     {
         $this->setName('run')
+            ->addOption('docker', null, InputOption::VALUE_OPTIONAL, 'Set this flag if EvoSC runs inside docker.', false)
             ->addOption('setup', null, InputOption::VALUE_OPTIONAL, 'Start the setup on boot.', false)
             ->addOption('skip_map_check', 'f', InputOption::VALUE_OPTIONAL, 'Start without verifying map integrity.',
                 false)
@@ -54,10 +62,16 @@ class EscRun extends Command
             ->setDescription('Run Evo Server Controller');
     }
 
+    /**
+     * @param InputInterface $input
+     * @param OutputInterface $output
+     */
     protected function initialize(InputInterface $input, OutputInterface $output)
     {
         global $serverName;
         global $__ManiaPlanet;
+        global $pidPath;
+        global $serverLogin;
 
         ConfigController::init();
         Log::setOutput($output);
@@ -84,6 +98,8 @@ class EscRun extends Command
             $_skipMapCheck = true;
         }
 
+        self::$docker = $input->getOption('docker') !== false;
+
         try {
             $output->writeln("Connecting to server...");
 
@@ -94,6 +110,14 @@ class EscRun extends Command
                 config('server.rpc.login'),
                 config('server.rpc.password')
             );
+
+            $serverLogin = Server::getSystemInfo()->serverLogin;
+            $pidPath = config('server.pidfile');
+
+            // if no config given, use original
+            if (empty($pidPath)) {
+                $pidPath = baseDir($serverLogin.'_evosc.pid');
+            }
 
             $serverName = Server::getServerName();
 
@@ -117,22 +141,28 @@ class EscRun extends Command
         }
     }
 
+    /**
+     * @param InputInterface $input
+     * @param OutputInterface $output
+     */
     protected function interact(InputInterface $input, OutputInterface $output)
     {
-      $pidPath = config('server.pidfile');
+        global $pidPath;
 
-      // if no config given, use original
-      if (empty($pidPath))
-        $pidPath = baseDir(config('server.login') . '_evosc.pid');
-
-      file_put_contents($pidPath, getmypid());
+        file_put_contents($pidPath, getmypid());
     }
 
+    /**
+     * @param InputInterface $input
+     * @param OutputInterface $output
+     * @return int
+     */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         global $__bootedVersion;
         global $_onlinePlayers;
         global $serverName;
+        global $serverLogin;
 
         $version = getEvoSCVersion();
         $motd = "      ______           _____ ______
@@ -170,7 +200,11 @@ class EscRun extends Command
         ModuleController::init();
         PlanetsController::init();
         CountdownController::init();
-        MatchController::init();
+        RoyalController::init();
+
+        EventController::init();
+        EventController::setServerLogin($serverLogin);
+
         ControllerController::loadControllers(Server::getScriptName()['CurrentValue'], true);
 
         self::addBootCommands();
@@ -192,7 +226,7 @@ class EscRun extends Command
         $map = Map::where('filename', Server::getCurrentMapInfo()->fileName)->first();
         Hook::fire('BeginMap', $map);
 
-        //Enable mode script rpc-callbacks else you wont get stuf flike checkpoints and finish
+        //Enable mode script rpc-callbacks else you wont get stuff like checkpoints and finish
         Server::triggerModeScriptEventArray('XmlRpc.EnableCallbacks', ['true']);
         Server::disableServiceAnnounces(true);
 
@@ -206,7 +240,7 @@ class EscRun extends Command
         $__bootedVersion = getEvoSCVersion();
 
         //cycle-loop
-        while (true) {
+        while ($this->keepRunning) {
             try {
                 Timer::startCycle();
                 RestClient::curlTick();
@@ -217,15 +251,15 @@ class EscRun extends Command
 
                 usleep($pause);
             } catch (Exception $e) {
-                Log::write('Failed to fetch callbacks from dedicated-server. Failed attempts: ' . $failedConnectionRequests . '/3');
-                Log::write($e->getMessage());
+                $message = 'Failed to fetch callbacks from dedicated-server. Failed attempts: ' . $failedConnectionRequests . '/3';
+                Log::errorWithCause($message, $e);
 
                 $failedConnectionRequests++;
                 if ($failedConnectionRequests > 3) {
                     Log::write('MPS',
                         sprintf('Connection terminated after %d connection-failures.', $failedConnectionRequests));
 
-                    return;
+                    return 1;
                 }
                 sleep(1);
             } catch (Error $e) {
@@ -235,26 +269,58 @@ class EscRun extends Command
                 $output->writeln("<error>===============================================================================</error>");
                 $output->writeln("<error>" . $e->getTraceAsString() . "</error>");
 
-                Log::write('EvoSC encountered an error: ' . $e->getMessage(), false);
-                Log::write($e->getTraceAsString(), false);
+                Log::errorWithCause("EvoSC encountered an error", $e, false);
             }
         }
+
+        return 0;
     }
 
+    /**
+     * Add EvoSC base commands
+     */
     public static function addBootCommands()
     {
         AwaitAction::createQueueAndStartCheckCycle();
 
-        AccessRight::add('restart_evosc', 'Allows you to restart EvoSC.');
+        if (!self::$docker) {
+            AccessRight::add('restart_evosc', 'Allows you to restart EvoSC.');
 
-        ChatCommand::add('//restart-evosc', function () {
-            restart_evosc();
-        }, 'Restart EvoSC', 'restart_evosc');
-
-        Timer::create('watch_for_restart_file', function () {
-            if (Cache::has('restart_evosc')) {
+            ChatCommand::add('//restart-evosc', function () {
                 restart_evosc();
-            }
-        }, '2m', true);
+            }, 'Restart EvoSC', 'restart_evosc');
+
+            Timer::create('watch_for_restart_file', function () {
+                if (Cache::has('restart_evosc')) {
+                    restart_evosc();
+                }
+            }, '30s', true);
+        }
+    }
+
+    /**
+     * Returns the signals which EvoSC subscribes to
+     *
+     * @return array
+     */
+    public function getSubscribedSignals(): array
+    {
+        // return here any of the constants defined by PCNTL extension
+        // https://www.php.net/manual/en/pcntl.constants.php
+        return [SIGTERM];
+    }
+
+    /**
+     * Signal handler
+     *
+     * @param int $signal
+     */
+    public function handleSignal(int $signal): void
+    {
+        if ($signal == SIGTERM) {
+            $this->keepRunning = false;
+            warningMessage('EvoSC received signal ', secondary('SIGTERM'), '. EvoSC Exiting.')->sendAdmin();
+            shutdown_evosc();
+        }
     }
 }
