@@ -22,6 +22,8 @@ use Exception;
  */
 class EventController implements ControllerInterface
 {
+    private static string $serverLogin;
+
     /**
      * Method called on controller-boot.
      */
@@ -115,9 +117,12 @@ class EventController implements ControllerInterface
             $player->spectator_status = $playerInfo['SpectatorStatus'];
             $player->player_id = $playerInfo['PlayerId'];
             $player->ubisoft_name = $playerInfo['NickName'];
+            $player->team = $playerInfo['TeamId'];
             if ($player->isDirty()) {
                 $player->save();
+                Hook::fire('PlayerInfoChanged', $player);
             }
+
             PlayerController::putPlayer($player);
         }
     }
@@ -133,11 +138,11 @@ class EventController implements ControllerInterface
             $login = $data[1];
             $text = $data[2];
 
-            if ($login === config('server.login')) {
+            if ($login === self::$serverLogin) {
                 return;
             }
 
-            $parts = explode(' ', $text);
+            $parts = explode(' ', trim($text));
 
             if (ChatCommand::has($parts[0])) {
                 ChatCommand::get($parts[0])->execute(player($login), $text);
@@ -161,13 +166,13 @@ class EventController implements ControllerInterface
             try {
                 Hook::fire('PlayerChat', player($login), $text);
             } catch (Exception $e) {
-                Log::write("Error: " . $e->getMessage());
+                Log::errorWithCause("Failed to fire PlayerChat hook", $e);
             }
 
             try {
                 ChatController::playerChat(player($login), $text);
             } catch (Exception $e) {
-                Log::write("Error: " . $e->getMessage());
+                Log::errorWithCause("Failed to send player text to chat", $e);
             }
         } else {
             throw new Exception('Malformed callback');
@@ -189,12 +194,13 @@ class EventController implements ControllerInterface
                     'NickName' => $details->nickName,
                     'path' => $details->path,
                     'player_id' => $details->playerId,
+                    'team' => $details->teamId
                 ]);
             } else {
-                $name = $details->nickName;
-
                 if ($name_ = Cache::get('nicknames/' . $playerInfo[0])) {
                     $name = $name_;
+                } else {
+                    $name = $details->nickName;
                 }
 
                 $player = Player::updateOrCreate(['Login' => $playerInfo[0]], [
@@ -202,6 +208,7 @@ class EventController implements ControllerInterface
                     'ubisoft_name' => $details->nickName,
                     'path' => $details->path,
                     'player_id' => $details->playerId,
+                    'team' => $details->teamId
                 ]);
 
                 $player->NickName = $name;
@@ -221,7 +228,11 @@ class EventController implements ControllerInterface
     private static function mpPlayerDisconnect($arguments)
     {
         if (count($arguments) == 2 && is_string($arguments[0])) {
-            Hook::fire('PlayerDisconnect', player($arguments[0]), 0);
+            $player = Player::updateOrCreate(['Login' => $arguments[0]], [
+                'player_id' => 0
+            ]);
+
+            Hook::fire('PlayerDisconnect', $player, 0);
         } else {
             throw new Exception('Malformed callback');
         }
@@ -248,7 +259,7 @@ class EventController implements ControllerInterface
             try {
                 Hook::fire('BeginMap', $map);
             } catch (Exception $e) {
-                Log::write("Error: " . $e->getMessage());
+                Log::errorWithCause("Failed to fire BeginMap hook", $e);
             }
         } else {
             throw new Exception('Malformed callback');
@@ -268,7 +279,7 @@ class EventController implements ControllerInterface
             try {
                 Hook::fire('EndMap', $map);
             } catch (Exception $e) {
-                Log::write("Error: " . $e->getMessage());
+                Log::errorWithCause("Failed to fire EndMap hook", $e);
             }
         } else {
             throw new Exception('Malformed callback');
@@ -286,7 +297,7 @@ class EventController implements ControllerInterface
             try {
                 ManiaLinkEvent::call(player($arguments[1]), $arguments[2], $arguments[3]);
             } catch (Exception $e) {
-                Log::write("Error: " . $e->getMessage());
+                Log::errorWithCause("Failed to call mania link", $e);
             }
         } else {
             throw new Exception('Malformed callback');
@@ -300,5 +311,10 @@ class EventController implements ControllerInterface
     {
         $file = cacheDir('round_start_time.txt');
         File::put($file, time());
+    }
+
+    public static function setServerLogin(string $serverLogin)
+    {
+        self::$serverLogin = $serverLogin;
     }
 }
