@@ -4,7 +4,9 @@
 namespace EvoSC\Controllers;
 
 
+use EvoSC\Classes\Cache;
 use EvoSC\Classes\ChatCommand;
+use EvoSC\Classes\Controller;
 use EvoSC\Classes\Hook;
 use EvoSC\Classes\Log;
 use EvoSC\Classes\Server;
@@ -12,7 +14,7 @@ use EvoSC\Classes\Template;
 use EvoSC\Interfaces\ControllerInterface;
 use EvoSC\Models\Player;
 
-class PointsController implements ControllerInterface
+class PointsController extends Controller implements ControllerInterface
 {
     /**
      * @var int
@@ -41,12 +43,12 @@ class PointsController implements ControllerInterface
             return;
         }
 
-        self::$originalPointsLimit = MatchSettingsController::getValueFromCurrentMatchSettings('S_PointsLimit');
-        if (self::$originalPointsLimit == -1) {
-            self::$originalPointsLimit = Server::getRoundPointsLimit()['CurrentValue'];
-        }
-        if (self::$currentPointsLimit == -1) {
-            self::$currentPointsLimit = Server::getRoundPointsLimit()['NextValue'];
+        if (Cache::has('points_controller')) {
+            list(self::$originalPointsLimit, self::$currentPointsLimit) = Cache::get('points_controller');
+            Cache::forget('points_controller');
+        } else {
+            self::$originalPointsLimit = (int)Server::getModeScriptSetting('S_PointsLimit');
+            self::$currentPointsLimit = (int)Server::getModeScriptSetting('S_PointsLimit');
         }
 
         Hook::add('PlayerConnect', [self::class, 'playerConnect']);
@@ -54,7 +56,13 @@ class PointsController implements ControllerInterface
 
         ChatCommand::add('//addpoints', [self::class, 'cmdAddPoints'], 'Add points to the points-limit.', 'manipulate_points');
 
-        Template::showAll('Helpers.update-points-limit', ['points' => self::$currentPointsLimit]);
+        Server::triggerModeScriptEventArray('Trackmania.GetPointsRepartition');
+        self::sendUpdatedPointsLimit(self::$currentPointsLimit);
+    }
+
+    public static function stop()
+    {
+        Cache::put('points_controller', [self::$originalPointsLimit, self::$currentPointsLimit], now()->addMinute());
     }
 
     /**
@@ -101,20 +109,15 @@ class PointsController implements ControllerInterface
         Server::setModeScriptSettings($modeScriptSettings);
         self::sendUpdatedPointsLimit(self::$originalPointsLimit);
         self::$currentPointsLimit = self::$originalPointsLimit;
+        Cache::forget('points_controller');
     }
 
     /**
      * @return int
      */
-    public static function getCurrentPoints(): int
+    public static function getCurrentPointsLimit(): int
     {
-        $modeScriptSettings = Server::getModeScriptSettings();
-
-        if (array_key_exists('S_PointsLimit', $modeScriptSettings)) {
-            return Server::getModeScriptSettings()['S_PointsLimit'];
-        }
-
-        return Server::getRoundPointsLimit()['CurrentValue'];
+        return self::$currentPointsLimit;
     }
 
     /**
@@ -134,21 +137,18 @@ class PointsController implements ControllerInterface
     }
 
     /**
-     * @return array
+     * @return int[]
      */
     public static function getPointsRepartition(): array
     {
-        $points = Server::getModeScriptSettings()['S_PointsRepartition'];
+        $repartition = Server::getModeScriptSetting('S_PointsRepartition');
 
-        //modescript: Trackmania.GetPointsRepartition
-
-        if ($points) {
-            $parts = explode(',', $points);
-            return array_map(function ($point) {
-                return intval($point);
-            }, $parts);
+        if (empty($repartition)) {
+            return [10, 9, 8, 7, 6, 5, 4, 3, 2, 1];
         }
 
-        return [10, 6, 4, 3, 2, 1];
+        return array_map(function ($i) {
+            return intval($i);
+        }, explode(',', $repartition));
     }
 }
