@@ -8,6 +8,7 @@ use EvoSC\Classes\Hook;
 use EvoSC\Classes\Module;
 use EvoSC\Classes\Template;
 use EvoSC\Controllers\MapController;
+use EvoSC\Controllers\ModeController;
 use EvoSC\Interfaces\ModuleInterface;
 use EvoSC\Models\Map;
 use EvoSC\Models\Player;
@@ -42,12 +43,14 @@ class CPRecords extends Module implements ModuleInterface
             self::$tracker = collect();
         }
 
-        $cpMode = config('cp-records.mode', 'best-cp');
-
-        if ($cpMode == 'best-round') {
-            Hook::add('PlayerFinish', [self::class, 'playerFinish']);
+        if (ModeController::isRoyal()) {
+            Hook::add('PlayerFinishSection', [self::class, 'playerFinishSection']);
         } else {
-            Hook::add('PlayerCheckpoint', [self::class, 'playerCheckpoint']);
+            if (config('cp-records.mode', 'best-cp') == 'best-round') {
+                Hook::add('PlayerFinish', [self::class, 'playerFinish']);
+            } else {
+                Hook::add('PlayerCheckpoint', [self::class, 'playerCheckpoint']);
+            }
         }
 
         Hook::add('PlayerConnect', [self::class, 'playerConnect']);
@@ -56,12 +59,35 @@ class CPRecords extends Module implements ModuleInterface
         Hook::add('BeginMatch', [self::class, 'beginMatch']);
     }
 
-    /**
-     *
-     */
     public function stop()
     {
         Cache::put('cp_records_current', self::$tracker->toArray(), now()->addMinute());
+    }
+
+    public static function playerFinishSection(Player $player, int $timeInMs, array $checkpoints, int $section)
+    {
+        if (!self::$tracker->has($section)) {
+            self::$tracker->put($section, (object)[
+                'section' => $section,
+                'time'    => $timeInMs,
+                'name'    => $player->NickName
+            ]);
+        } else {
+            $tracker = self::$tracker->get($section);
+
+            if ($timeInMs >= $tracker->time) {
+                return;
+            }
+
+            self::$tracker->put($section, (object)[
+                'section' => $section,
+                'time'    => $timeInMs,
+                'name'    => $player->NickName
+            ]);
+        }
+
+        $data = self::$tracker->values()->toJson();
+        Template::showAll('CPRecords.update_royal', compact('data', 'section'));
     }
 
     /**
@@ -137,11 +163,19 @@ class CPRecords extends Module implements ModuleInterface
 
     /**
      * @param Player $player
+     * @throws \EvoSC\Exceptions\InvalidArgumentException
      */
     public static function playerConnect(Player $player)
     {
-        self::sendUpdatedCpRecords();
-        Template::show($player, 'CPRecords.widget');
+        if (ModeController::isRoyal()) {
+            $data = self::$tracker->values()->toJson();
+            $section = -1;
+            Template::showAll('CPRecords.update_royal', compact('data', 'section'));
+            Template::show($player, 'CPRecords.widget_royal');
+        } else {
+            self::sendUpdatedCpRecords();
+            Template::show($player, 'CPRecords.widget');
+        }
     }
 
     /**
@@ -156,9 +190,16 @@ class CPRecords extends Module implements ModuleInterface
     /**
      *
      */
-    public static function beginMatch()
-    {
+    public static function beginMatch(){
         self::$tracker = collect();
-        Template::showAll('CPRecords.widget');
+        if (ModeController::isRoyal()) {
+            $data = self::$tracker->values()->toJson();
+            $section = -1;
+            Template::showAll('CPRecords.update_royal', compact('data', 'section'));
+            Template::showAll('CPRecords.widget_royal');
+        }else{
+            self::sendUpdatedCpRecords();
+            Template::showAll('CPRecords.widget');
+        }
     }
 }
