@@ -16,6 +16,7 @@ use EvoSC\Interfaces\ModuleInterface;
 use EvoSC\Models\AccessRight;
 use EvoSC\Models\Map;
 use EvoSC\Models\Player;
+use EvoSC\Models\Schedule;
 use EvoSC\Modules\QuickButtons\QuickButtons;
 use SimpleXMLElement;
 
@@ -26,38 +27,38 @@ class MatchSettingsManager extends Module implements ModuleInterface
 
     private static array $modeTemplatesManiaplanet = [
         'TimeAttack' => 'timeattack.xml',
-        'Team' => 'team.xml',
-        'Rounds' => 'rounds.xml',
-        'Laps' => 'laps.xml',
-        'Cup' => 'cup.xml',
-        'Chase' => 'chase.xml',
+        'Team'       => 'team.xml',
+        'Rounds'     => 'rounds.xml',
+        'Laps'       => 'laps.xml',
+        'Cup'        => 'cup.xml',
+        'Chase'      => 'chase.xml',
     ];
 
     private static array $modeTemplatesTrackmania = [
         'TimeAttack' => 'timeattack.xml',
-        'Team' => 'team.xml',
-        'Rounds' => 'rounds.xml',
-        'Laps' => 'laps.xml',
-        'Cup' => 'cup.xml',
+        'Team'       => 'team.xml',
+        'Rounds'     => 'rounds.xml',
+        'Laps'       => 'laps.xml',
+        'Cup'        => 'cup.xml',
     ];
 
     private static array $gameModesManiaplanet = [
         'TimeAttack' => 'TimeAttack.Script.txt',
-        'Rounds' => 'Rounds.Script.txt',
-        'Team' => 'Team.Script.txt',
-        'Cup' => 'Cup.Script.txt',
-        'Laps' => 'Laps.Script.txt',
-        'Chase' => 'Chase.Script.txt',
+        'Rounds'     => 'Rounds.Script.txt',
+        'Team'       => 'Team.Script.txt',
+        'Cup'        => 'Cup.Script.txt',
+        'Laps'       => 'Laps.Script.txt',
+        'Chase'      => 'Chase.Script.txt',
     ];
 
     private static array $gameModesTrackmania = [
         'TimeAttack' => 'Trackmania/TM_TimeAttack_Online.Script.txt',
-        'Rounds' => 'Trackmania/TM_Rounds_Online.Script.txt',
-        'Teams' => 'Trackmania/TM_Teams_Online.Script.txt',
-        'Cup' => 'Trackmania/TM_Cup_Online.Script.txt',
-        'Laps' => 'Trackmania/TM_Laps_Online.Script.txt',
-        'Champion' => 'Trackmania/TM_Champion_Online.Script.txt',
-        'Knockout' => 'Trackmania/TM_Knockout_Online.Script.txt',
+        'Rounds'     => 'Trackmania/TM_Rounds_Online.Script.txt',
+        'Teams'      => 'Trackmania/TM_Teams_Online.Script.txt',
+        'Cup'        => 'Trackmania/TM_Cup_Online.Script.txt',
+        'Laps'       => 'Trackmania/TM_Laps_Online.Script.txt',
+        'Champion'   => 'Trackmania/TM_Champion_Online.Script.txt',
+        'Knockout'   => 'Trackmania/TM_Knockout_Online.Script.txt',
     ];
 
     /**
@@ -76,25 +77,40 @@ class MatchSettingsManager extends Module implements ModuleInterface
 
         ChatCommand::add('//msm', [self::class, 'showOverview'], 'Show MatchSettingsManager', 'matchsettings_edit');
 
-        ManiaLinkEvent::add('msm.overview', [self::class, 'showOverview'], 'matchsettings_edit');
+        ManiaLinkEvent::add('msm.load', [self::class, 'loadMatchsettings'], 'matchsettings_load');
+        ManiaLinkEvent::add('msm.load_and_skip', [self::class, 'loadMatchsettingsAndSkip'], 'matchsettings_load');
+        ManiaLinkEvent::add('msm.overview', [self::class, 'showOverview'], 'matchsettings_load');
         ManiaLinkEvent::add('msm.create', [self::class, 'showCreateMatchsettings'], 'matchsettings_edit');
         ManiaLinkEvent::add('msm.edit', [self::class, 'showEditMatchsettings'], 'matchsettings_edit');
         ManiaLinkEvent::add('msm.edit_maps', [self::class, 'showEditMatchsettingsMaps'], 'matchsettings_edit');
-        ManiaLinkEvent::add('msm.load', [self::class, 'loadMatchsettings'], 'matchsettings_edit');
+        ManiaLinkEvent::add('msm.edit_folders', [self::class, 'showEditMatchsettingsFolders'], 'matchsettings_edit');
         ManiaLinkEvent::add('msm.duplicate', [self::class, 'duplicateMatchsettings'], 'matchsettings_edit');
         ManiaLinkEvent::add('msm.delete', [self::class, 'deleteMatchsettings'], 'matchsettings_edit');
         ManiaLinkEvent::add('msm.new', [self::class, 'createNewMatchsettings'], 'matchsettings_edit');
         ManiaLinkEvent::add('msm.update', [self::class, 'updateMatchsettings'], 'matchsettings_edit');
         ManiaLinkEvent::add('msm.save_maps', [self::class, 'saveMaps'], 'matchsettings_edit');
+        ManiaLinkEvent::add('msm.save_folders', [self::class, 'saveFolders'], 'matchsettings_edit');
+        ManiaLinkEvent::add('msm.schedule', [self::class, 'scheduleMatchSettings'], 'matchsettings_load');
 
         if (config('quick-buttons.enabled')) {
-            QuickButtons::addButton('', 'MatchSetting Manager', 'msm.overview', 'map.edit');
+            QuickButtons::addButton('', 'MatchSetting Manager', 'msm.overview', 'matchsettings_load');
         }
     }
 
     /**
      * @param Player $player
+     * @param $timeStamp
+     * @param $matchsettingsFile
+     */
+    public static function scheduleMatchSettings(Player $player, $timeStamp, $matchsettingsFile)
+    {
+        Schedule::maniaLinkEvent($player, 'Load ' . $matchsettingsFile, $timeStamp, 'msm.load_and_skip', serverPlayer(), $matchsettingsFile);
+    }
+
+    /**
+     * @param Player $player
      * @param $data
+     * @throws \EvoSC\Exceptions\InvalidArgumentException
      */
     public static function saveMaps(Player $player, $data)
     {
@@ -123,6 +139,41 @@ class MatchSettingsManager extends Module implements ModuleInterface
 
         successMessage('MatchSettings ', secondary($matchsettings), ' saved.')->send($player);
         self::showEditMatchsettingsMaps($player, $matchsettings);
+    }
+
+    /**
+     * @param Player $player
+     * @param $data
+     * @throws \EvoSC\Exceptions\InvalidArgumentException
+     */
+    public static function saveFolders(Player $player, $data)
+    {
+        $matchsettings = $data->matchsettings;
+
+        $enabledFolders = collect($data)->map(function ($value, $key) {
+            if (preg_match('/^folder_.+/', $key) && $value == '1') {
+                return preg_replace('/^folder_/', '', $key);
+            }
+            return null;
+        })->filter()->values();
+
+        $file = mapsDir("MatchSettings/$matchsettings.txt");
+
+        $settings = new SimpleXMLElement(File::get($file));
+        unset($settings->map);
+
+        $test = Map::whereIn('folder', $enabledFolders)->get();
+
+        foreach ($test as $map) {
+            $entry = $settings->addChild('map');
+            $entry->addChild('file', $map->filename);
+            $entry->addChild('ident', $map->uid);
+        }
+
+        File::put($file, Utility::simpleXmlPrettyPrint($settings));
+
+        successMessage('MatchSettings ', secondary($matchsettings), ' saved.')->send($player);
+        self::showEditMatchsettingsFolders($player, $matchsettings);
     }
 
     /**
@@ -213,6 +264,41 @@ class MatchSettingsManager extends Module implements ModuleInterface
 
     /**
      * @param Player $player
+     * @param string $name
+     * @throws \EvoSC\Exceptions\InvalidArgumentException
+     */
+    public static function showEditMatchsettingsFolders(Player $player, string $name)
+    {
+        $file = Server::getMapsDirectory() . 'MatchSettings/' . $name . '.txt';
+        $data = File::get($file);
+        $enabledFolders = collect();
+        $maps = collect();
+        $xml = new SimpleXMLElement($data);
+
+        foreach ($xml as $node) {
+            if ($node->getName() == 'map') {
+                $value = (string)$node->ident;
+                $maps->push($value);
+            }
+        }
+
+        $folders = Map::distinct()->get(['folder']);
+
+        foreach ($folders as $folder) {
+            $uid_mapper = function ($item) {return $item->uid;};
+            $count = Map::whereFolder($folder->folder)->get()->toBase()->map($uid_mapper)->diff($maps->toBase())->count();
+            if($count == 0) {
+                $enabledFolders->push($folder->folder);
+            }
+        }
+
+        $folderChunks = $folders->chunk(19);
+
+        Template::show($player, 'MatchSettingsManager.edit-folders', compact('name', 'folderChunks', 'enabledFolders'));
+    }
+
+    /**
+     * @param Player $player
      * @param string $oldFilename
      * @param string $filename
      * @param mixed ...$settings
@@ -260,8 +346,6 @@ class MatchSettingsManager extends Module implements ModuleInterface
         } else {
             $scriptName = self::$gameModesTrackmania[$modeName];
         }
-
-        dump($modeName, $scriptName);
 
         $content = File::get($sourceMatchsettings);
         $content = str_replace('%script_name%', $scriptName, $content);
@@ -315,6 +399,16 @@ class MatchSettingsManager extends Module implements ModuleInterface
     public static function loadMatchsettings(Player $player, string $matchsettingsFile)
     {
         MatchSettingsController::loadMatchSettings(true, $player, $matchsettingsFile . '.txt');
+    }
+
+    /**
+     * @param Player $player
+     * @param string $matchSettingsFile
+     */
+    public static function loadMatchsettingsAndSkip(Player $player, string $matchSettingsFile)
+    {
+        self::loadMatchsettings($player, $matchSettingsFile);
+        Server::nextMap();
     }
 
     /**
