@@ -71,7 +71,7 @@ class MxDownload extends Module implements ModuleInterface
                 return;
             }
         } catch (Exception $e) {
-            Log::error($e->getMessage());
+            Log::errorWithCause("The map $tmxId is unknown", $e);
             dangerMessage('The map with the id ', secondary($tmxId), ' is unknown.')->send($player);
             return;
         }
@@ -80,7 +80,7 @@ class MxDownload extends Module implements ModuleInterface
             $comment = self::parseBB($details->Comments);
             Template::show($player, 'MxDownload.add-map-info', compact('details', 'comment'));
         } catch (Exception $e) {
-            Log::error($e->getMessage());
+            Log::errorWithCause('Failed to parse map info', $e);
             warningMessage('Failed to parse map info, adding...')->send($player);
             self::addMap($player, $tmxId);
         }
@@ -113,7 +113,7 @@ class MxDownload extends Module implements ModuleInterface
         $filename = html_entity_decode(trim($filename), ENT_QUOTES | ENT_HTML5);
         $filename = preg_replace('/[^a-z0-9\-_# .]/i', '', $filename);
         $filename = preg_replace('/\s/i', '_', $filename);
-        $filename = "MX/$mxId" . "_$filename";
+        $filename = "MX" . DIRECTORY_SEPARATOR . "$mxId" . "_$filename";
 
         Log::write('Saving map as ' . MapController::getMapsPath($filename), true);
         File::put(MapController::getMapsPath($filename), $download->getBody()->getContents());
@@ -135,7 +135,7 @@ class MxDownload extends Module implements ModuleInterface
     public static function addMap(Player $player, int $mxId)
     {
         if ($mxId <= 0) {
-            warningMessage('MX-ID invalid.')->send($player);
+            warningMessage('Exchange-ID invalid.')->send($player);
             return;
         }
 
@@ -152,7 +152,7 @@ class MxDownload extends Module implements ModuleInterface
 
         $mxDetails = self::loadMxDetails($mxId);
         Log::write(json_encode($mxDetails));
-        $gbx = json_decode(MapController::getGbxInformation($filename, true));
+        $gbx = MapController::getGbxInformation($filename);
         Log::write(json_encode($gbx));
 
         if (!isset($gbx->MapUid)) {
@@ -176,6 +176,7 @@ class MxDownload extends Module implements ModuleInterface
             'name' => $gbx->Name,
             'environment' => $gbx->Environment,
             'title_id' => $gbx->TitleId,
+            'folder' => substr($filename, 0, strrpos($filename, DIRECTORY_SEPARATOR)),
             'enabled' => 1,
             'cooldown' => config('server.map-cooldown', 10),
             'mx_id' => $mxId,
@@ -187,8 +188,8 @@ class MxDownload extends Module implements ModuleInterface
                 Server::addMap($filename);
                 Log::info("Added $filename to the selection.");
             } catch (Exception $e) {
+                Log::errorWithCause('Adding map to selection failed', $e);
                 warningMessage('Failed to add map ', secondary($gbx->Name), ' to the map-pool: ' . $e->getMessage())->send($player);
-                Log::write('Adding map to selection failed: ' . $e->getMessage());
                 return;
             }
         }
